move sims for tests into sim directory. tweaks to nsmb fit
