+++ qetpy/sim/__init__.py
@@ -1,1 +1,2 @@
 from ._sim import *
+from ._sim_nsmb import *
